+++ web_src/src/pages/workflowv2/lib/canvas-groups.ts
@@ -2,7 +2,8 @@ import type { CanvasesCanvas, ComponentsNode } from "@/api-client";
 import { DefaultLayoutEngine } from "@/lib/layout";
 import type { CanvasNode } from "@/ui/CanvasPage";
 import { GROUP_CHILD_EDGE_PADDING, GROUP_CHILD_MIN_Y_OFFSET, normalizeGroupColor } from "@/ui/groupNode/constants";
-import { buildChildToGroupMap, collectGroupChildIds, generateNodeId, generateUniqueNodeName } from "../utils";
+import { generateNodeId, generateUniqueNodeName } from "../utils";
+import { buildChildToGroupMap, collectGroupChildIds } from "@/lib/canvas/groups";
 
 const DEFAULT_GROUP_WIDTH = 480;
 const DEFAULT_GROUP_HEIGHT = 320;
